Fixed minor error on import of zipped project with old (*.neuro.xml) file extension
git-svn-id: svn://neuroconstruct.org:3999/neuroConstruct/trunk@1051 cb8a3d95-5132-0410-b650-8d485996f008

+++ src/ucl/physiol/neuroconstruct/gui/MainFrame.java
@@ -7451,9 +7451,23 @@ public class MainFrame extends JFrame implements ProjectEventListener, Generatio
 
                             if (!(new File(nameOfNewProjectFile)).exists())
                             {
-                                GuiUtils.showErrorMessage(logger, "The expected project file: "+ nameOfNewProjectFile + " was not found!", null, this);
-                                projDirToCreate.delete();
-                                return;
+
+                                String nameOfOldProjectFile = projDirToCreate.getAbsolutePath()
+                                                     + System.getProperty("file.separator")
+                                                     + newProjName
+                                                     + ProjectStructure.getOldProjectFileExtension();
+
+                                if (!(new File(nameOfOldProjectFile)).exists())
+                                {
+                                    GuiUtils.showErrorMessage(logger, "The expected project file: "+ nameOfNewProjectFile
+                                            + " was not found (neither was: "+nameOfOldProjectFile+")! ", null, this);
+                                    projDirToCreate.delete();
+                                    return;
+                                }
+                                else
+                                {
+                                    nameOfNewProjectFile = nameOfOldProjectFile;
+                                }
                             }
 
                             doLoadProject(nameOfNewProjectFile);
